requirements.txt
# Django
Django==6.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11

# AI / LLM
openai==1.83.0
crewai==1.9.3
crewai-tools==1.9.3
chromadb==1.1.1

# WhatsApp / Twilio
twilio==9.10.1
requests==2.32.5

# Configuration
python-decouple==3.8
